Remove old initiation end point, services and tests
