+++ src/test/java/uk/gov/hmcts/reform/wataskmanagementapi/controllers/ExclusiveTaskActionsControllerTest.java
@@ -10,10 +10,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import uk.gov.hmcts.reform.wataskmanagementapi.auth.restrict.ClientAccessControlService;
 import uk.gov.hmcts.reform.wataskmanagementapi.cft.entities.TaskResource;
-import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.InitiateTaskRequestAttributes;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.InitiateTaskRequestMap;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.TerminateTaskRequest;
-import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.entities.TaskAttribute;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.options.TerminateInfo;
 import uk.gov.hmcts.reform.wataskmanagementapi.exceptions.v2.GenericForbiddenException;
 import uk.gov.hmcts.reform.wataskmanagementapi.services.TaskManagementService;
@@ -21,7 +19,6 @@ import uk.gov.hmcts.reform.wataskmanagementapi.services.TaskManagementService;
 import java.util.Map;
 import java.util.UUID;
 
-import static java.util.Arrays.asList;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -94,60 +91,6 @@ class ExclusiveTaskActionsControllerTest {
                 Map.of(
                     TASK_TITLE.value(), "aTaskTitle",
                     TASK_NAME.value(), "aTaskName"
-                )
-            );
-
-            when(clientAccessControlService.hasExclusiveAccess(SERVICE_AUTHORIZATION_TOKEN))
-                .thenReturn(false);
-
-            assertThatThrownBy(() -> exclusiveTaskActionsController.initiate(SERVICE_AUTHORIZATION_TOKEN, taskId, req))
-                .isInstanceOf(GenericForbiddenException.class)
-                .hasNoCause()
-                .hasMessage("Forbidden: "
-                                + "The action could not be completed because the "
-                                + "client/user had insufficient rights to a resource.");
-        }
-
-    }
-
-    @Nested
-    class OldInitiateRequest {
-
-        @Test
-        void should_succeed_when_initiating_a_task_and_return_201() {
-
-            InitiateTaskRequestAttributes req = new InitiateTaskRequestAttributes(
-                INITIATION,
-                asList(
-                    new TaskAttribute(TASK_TYPE, "followUpOverdueReasonsForAppeal"),
-                    new TaskAttribute(TASK_NAME, "follow Up Overdue Reasons For Appeal")
-                )
-            );
-
-            when(clientAccessControlService.hasExclusiveAccess(SERVICE_AUTHORIZATION_TOKEN))
-                .thenReturn(true);
-            when(taskManagementService.initiateTask(taskId, req))
-                .thenReturn(createDummyTaskResource(taskId));
-            ResponseEntity<TaskResource> response = exclusiveTaskActionsController
-                .initiate(SERVICE_AUTHORIZATION_TOKEN, taskId, req);
-
-            assertNotNull(response);
-            assertEquals(HttpStatus.CREATED, response.getStatusCode());
-            assertNotNull(response.getBody());
-            assertEquals(taskId, response.getBody().getTaskId());
-            assertEquals("follow Up Overdue Reasons For Appeal", response.getBody().getTaskName());
-            assertEquals("followUpOverdueReasonsForAppeal", response.getBody().getTaskType());
-            assertEquals(UNCONFIGURED, response.getBody().getState());
-        }
-
-        @Test
-        void should_fail_when_initiating_a_task_and_client_not_whitelisted_and_return_403() {
-
-            InitiateTaskRequestAttributes req = new InitiateTaskRequestAttributes(
-                INITIATION,
-                asList(
-                    new TaskAttribute(TASK_TITLE, "aTaskTitle"),
-                    new TaskAttribute(TASK_NAME, "aTaskName")
                 )
             );
