+++ src/test/java/uk/gov/hmcts/reform/wataskmanagementapi/services/CFTTaskMapperTest.java
@@ -20,7 +20,6 @@ import uk.gov.hmcts.reform.wataskmanagementapi.cft.enums.BusinessContext;
 import uk.gov.hmcts.reform.wataskmanagementapi.cft.enums.CFTTaskState;
 import uk.gov.hmcts.reform.wataskmanagementapi.cft.enums.ExecutionType;
 import uk.gov.hmcts.reform.wataskmanagementapi.cft.enums.TaskSystem;
-import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.entities.TaskAttribute;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition;
 import uk.gov.hmcts.reform.wataskmanagementapi.data.RoleAssignmentCreator;
 import uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition;
@@ -61,7 +60,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static uk.gov.hmcts.reform.wataskmanagementapi.cft.enums.CFTTaskState.COMPLETED;
 import static uk.gov.hmcts.reform.wataskmanagementapi.cft.enums.CFTTaskState.UNCONFIGURED;
 import static uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition.TASK_CASE_ID;
-import static uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition.TASK_DUE_DATE;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaTime.CAMUNDA_DATA_TIME_FORMATTER;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaValue.booleanValue;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaValue.integerValue;
@@ -69,13 +67,9 @@ import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.Ca
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.ADDITIONAL_PROPERTIES;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.ASSIGNEE;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.AUTO_ASSIGNED;
-import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.CASE_CATEGORY;
-import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.CASE_ID;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.CASE_MANAGEMENT_CATEGORY;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.CASE_NAME;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.CASE_TYPE_ID;
-import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.CREATED;
-import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.DESCRIPTION;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.DUE_DATE;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.EXECUTION_TYPE;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.HAS_WARNINGS;
@@ -119,14 +113,13 @@ class CFTTaskMapperTest {
 
     @Test
     void given_null_attribute_when_mapToTaskResource_then_dont_throw_exception() {
-        List<TaskAttribute> attributes = new ArrayList<>();
+        Map<String, Object> attributes = new HashMap<>();
         ZonedDateTime createdDate = ZonedDateTime.now();
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        attributes.add(new TaskAttribute(TaskAttributeDefinition.TASK_ASSIGNEE, "someAssignee"));
-        attributes.add(new TaskAttribute(TASK_DUE_DATE, formattedDueDate));
-        attributes.add(null);
+        attributes.put(TaskAttributeDefinition.TASK_ASSIGNEE.value(), "someAssignee");
+        attributes.put(DUE_DATE.value(), formattedDueDate);
         assertDoesNotThrow(() -> {
             cftTaskMapper.mapToTaskResource(taskId, attributes);
         });
@@ -140,7 +133,7 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate, null);
+        Map<String, Object> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate, null);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
 
@@ -194,77 +187,17 @@ class CFTTaskMapperTest {
     }
 
     @Test
-    void should_map_initiation_attributes_to_cft_task() {
-        ZonedDateTime createdDate = ZonedDateTime.now();
-        String formattedCreatedDate = CAMUNDA_DATA_TIME_FORMATTER.format(createdDate);
-        ZonedDateTime dueDate = createdDate.plusDays(1);
-        String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
-
-        Map<String, Object> attributes
-            = getDefaultInitiationAttributes(formattedCreatedDate, formattedDueDate, formattedDueDate);
-
-        TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
-
-        assertEquals("SOME_TASK_ID", taskResource.getTaskId());
-        assertEquals("someCamundaTaskName", taskResource.getTaskName());
-        assertEquals("someTaskType", taskResource.getTaskType());
-        assertEquals(
-            OffsetDateTime.parse(formattedDueDate, CAMUNDA_DATA_TIME_FORMATTER),
-            taskResource.getDueDateTime()
-        );
-        assertEquals(CFTTaskState.UNCONFIGURED, taskResource.getState());
-        assertEquals(TaskSystem.SELF, taskResource.getTaskSystem());
-        assertEquals(SecurityClassification.PUBLIC, taskResource.getSecurityClassification());
-        assertEquals("someTitle", taskResource.getTitle());
-        assertEquals("someCamundaTaskDescription", taskResource.getDescription());
-        assertNull(taskResource.getNotes());
-        assertEquals(5000, taskResource.getMajorPriority());
-        assertEquals(500, taskResource.getMinorPriority());
-        assertEquals("someAssignee", taskResource.getAssignee());
-        assertEquals(false, taskResource.getAutoAssigned());
-        assertNull(taskResource.getWorkTypeResource());
-        assertNull(taskResource.getRoleCategory());
-        assertEquals(true, taskResource.getHasWarnings());
-        assertThat(taskResource.getNotes()).isNull();
-        assertNull(taskResource.getAssignmentExpiry());
-        assertEquals("00000", taskResource.getCaseId());
-        assertEquals("someCaseType", taskResource.getCaseTypeId());
-        assertEquals("someCaseName", taskResource.getCaseName());
-        assertEquals("someJurisdiction", taskResource.getJurisdiction());
-        assertEquals("someRegion", taskResource.getRegion());
-        assertNull(taskResource.getRegionName());
-        assertEquals("someStaffLocationId", taskResource.getLocation());
-        assertEquals("someStaffLocationName", taskResource.getLocationName());
-        assertEquals(EXPECTED_ADDITIONAL_PROPERTIES, taskResource.getAdditionalProperties());
-        assertNull(taskResource.getBusinessContext());
-        assertNull(taskResource.getTerminationReason());
-        assertEquals(
-            OffsetDateTime.parse(formattedCreatedDate, CAMUNDA_DATA_TIME_FORMATTER),
-            taskResource.getCreated()
-        );
-        assertEquals(new ExecutionTypeResource(
-            ExecutionType.MANUAL,
-            ExecutionType.MANUAL.getName(),
-            ExecutionType.MANUAL.getDescription()
-        ), taskResource.getExecutionTypeCode());
-        assertNull(taskResource.getTaskRoleResources());
-        assertEquals(
-            OffsetDateTime.parse(formattedDueDate, CAMUNDA_DATA_TIME_FORMATTER),
-            taskResource.getPriorityDate()
-        );
-    }
-
-
-    @Test
     void should_map_task_attributes_to_cft_task_with_warnings() {
         ZonedDateTime createdDate = ZonedDateTime.now();
         String formattedCreatedDate = CAMUNDA_DATA_TIME_FORMATTER.format(createdDate);
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes = getDefaultAttributesWithWarnings(formattedCreatedDate, formattedDueDate);
+
+        Map<String, Object> attributes = getDefaultAttributesWithWarnings(formattedCreatedDate, formattedDueDate);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
+
 
         assertEquals("SOME_TASK_ID", taskResource.getTaskId());
         assertEquals("someCamundaTaskName", taskResource.getTaskName());
@@ -327,7 +260,7 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes = getDefaultAttributesWithWarnings(formattedCreatedDate, formattedDueDate);
+        Map<String, Object> attributes = getDefaultAttributesWithWarnings(formattedCreatedDate, formattedDueDate);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
         Set<PermissionTypes> permissionsUnion = new HashSet<>(
@@ -341,6 +274,7 @@ class CFTTaskMapperTest {
             )
         );
         Task task = cftTaskMapper.mapToTaskWithPermissions(taskResource, permissionsUnion);
+
 
         assertNotNull(task);
         assertTrue(task.getWarnings());
@@ -549,8 +483,7 @@ class CFTTaskMapperTest {
             ExecutionType.MANUAL.getDescription()
         ), taskResource.getExecutionTypeCode());
         assertNotNull(taskResource.getTaskRoleResources());
-        List<TaskRoleResource> roleResourcesList = new ArrayList<>();
-        roleResourcesList.addAll(taskResource.getTaskRoleResources());
+        List<TaskRoleResource> roleResourcesList = new ArrayList<>(taskResource.getTaskRoleResources());
 
         assertEquals("senior-tribunal-caseworker", roleResourcesList.get(0).getRoleName());
         assertEquals(true, roleResourcesList.get(0).getRead());
@@ -559,16 +492,6 @@ class CFTTaskMapperTest {
         assertEquals(true, roleResourcesList.get(0).getManage());
         assertEquals(true, roleResourcesList.get(0).getCancel());
         assertEquals(true, roleResourcesList.get(0).getRefer());
-        assertEquals(false, roleResourcesList.get(0).getClaim());
-        assertEquals(false, roleResourcesList.get(0).getAssign());
-        assertEquals(false, roleResourcesList.get(0).getUnassign());
-        assertEquals(false, roleResourcesList.get(0).getUnassignAssign());
-        assertEquals(false, roleResourcesList.get(0).getComplete());
-        assertEquals(false, roleResourcesList.get(0).getCompleteOwn());
-        assertEquals(false, roleResourcesList.get(0).getCancelOwn());
-        assertEquals(false, roleResourcesList.get(0).getUnassignClaim());
-        assertEquals(false, roleResourcesList.get(0).getUnclaim());
-        assertEquals(false, roleResourcesList.get(0).getUnclaimAssign());
         assertArrayEquals(new String[]{}, roleResourcesList.get(0).getAuthorizations());
         assertEquals("tribunal-caseworker", roleResourcesList.get(1).getRoleName());
         assertEquals(true, roleResourcesList.get(1).getRead());
@@ -577,16 +500,6 @@ class CFTTaskMapperTest {
         assertEquals(true, roleResourcesList.get(1).getManage());
         assertEquals(true, roleResourcesList.get(1).getCancel());
         assertEquals(true, roleResourcesList.get(1).getRefer());
-        assertEquals(false, roleResourcesList.get(0).getClaim());
-        assertEquals(false, roleResourcesList.get(0).getAssign());
-        assertEquals(false, roleResourcesList.get(0).getUnassign());
-        assertEquals(false, roleResourcesList.get(0).getUnassignAssign());
-        assertEquals(false, roleResourcesList.get(0).getComplete());
-        assertEquals(false, roleResourcesList.get(0).getCompleteOwn());
-        assertEquals(false, roleResourcesList.get(0).getCancelOwn());
-        assertEquals(false, roleResourcesList.get(0).getUnassignClaim());
-        assertEquals(false, roleResourcesList.get(0).getUnclaim());
-        assertEquals(false, roleResourcesList.get(0).getUnclaimAssign());
         assertArrayEquals(new String[]{"IA", "WA"}, roleResourcesList.get(1).getAuthorizations());
         assertNull(taskResource.getNextHearingId());
         assertNull(taskResource.getNextHearingDate());
@@ -686,9 +599,8 @@ class CFTTaskMapperTest {
     @Test
     void should_throw_exception_when_execution_type_enum_is_not_mapped() {
 
-        List<TaskAttribute> attributes = singletonList(
-            new TaskAttribute(TaskAttributeDefinition.TASK_EXECUTION_TYPE_NAME, "someExecutionType")
-        );
+        Map<String, Object> attributes = Map.of(
+            CamundaVariableDefinition.EXECUTION_TYPE.value(), "someExecutionType");
 
         assertThatThrownBy(() -> cftTaskMapper.mapToTaskResource(taskId, attributes))
             .isInstanceOf(IllegalStateException.class)
@@ -699,15 +611,15 @@ class CFTTaskMapperTest {
     @Test
     void should_throw_exception_when_no_due_date() {
 
-        List<TaskAttribute> attributes = asList(
-            new TaskAttribute(TaskAttributeDefinition.TASK_TYPE, "aTaskType"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NAME, "aTaskName"),
-            new TaskAttribute(TASK_CASE_ID, "someCaseId")
+        Map<String, Object> attributes = Map.of(
+            CamundaVariableDefinition.TASK_TYPE.value(), "aTaskType",
+            CamundaVariableDefinition.TASK_NAME.value(), "aTaskName",
+            TASK_CASE_ID.value(), "someCaseId"
         );
 
         assertThatThrownBy(() -> cftTaskMapper.mapToTaskResource(taskId, attributes))
             .isInstanceOf(NullPointerException.class)
-            .hasMessage("TASK_DUE_DATE must not be null")
+            .hasMessage("DUE_DATE must not be null")
             .hasNoCause();
     }
 
@@ -717,18 +629,17 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes = asList(
-            new TaskAttribute(TaskAttributeDefinition.TASK_EXECUTION_TYPE_NAME, "MANUAL"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SYSTEM, "someTaskSystem"),
-            new TaskAttribute(TASK_DUE_DATE, formattedDueDate)
-
+        Map<String, Object> attributes = Map.of(
+            CamundaVariableDefinition.EXECUTION_TYPE.value(), "MANUAL",
+            CamundaVariableDefinition.TASK_SYSTEM.value(), "someTaskSystem",
+            DUE_DATE.value(), formattedDueDate
         );
 
         assertThatThrownBy(() -> cftTaskMapper.mapToTaskResource(taskId, attributes))
             .isInstanceOf(IllegalArgumentException.class)
             .hasMessageContaining(
                 "Cannot deserialize value of type `uk.gov.hmcts.reform.wataskmanagementapi.cft.enums.TaskSystem` "
-                + "from String \"someTaskSystem\": not one of the values accepted for Enum class: [CTSC, SELF]")
+                    + "from String \"someTaskSystem\": not one of the values accepted for Enum class: [CTSC, SELF]")
             .hasCauseInstanceOf(InvalidFormatException.class);
 
     }
@@ -739,20 +650,20 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes = asList(
-            new TaskAttribute(TaskAttributeDefinition.TASK_EXECUTION_TYPE_NAME, "MANUAL"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SYSTEM, "SELF"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SECURITY_CLASSIFICATION, "someInvalidEnumValue"),
-            new TaskAttribute(TASK_DUE_DATE, formattedDueDate)
+        Map<String, Object> attributes = Map.of(
+            CamundaVariableDefinition.EXECUTION_TYPE.value(), "MANUAL",
+            CamundaVariableDefinition.TASK_SYSTEM.value(), "SELF",
+            CamundaVariableDefinition.SECURITY_CLASSIFICATION.value(), "someInvalidEnumValue",
+            DUE_DATE.value(), formattedDueDate
         );
 
         assertThatThrownBy(() -> cftTaskMapper.mapToTaskResource(taskId, attributes))
             .isInstanceOf(IllegalArgumentException.class)
             .hasMessageContaining(
                 "Cannot deserialize value of type "
-                + "`uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.SecurityClassification` "
-                + "from String \"someInvalidEnumValue\": not one of the values accepted for Enum class: "
-                + "[PUBLIC, RESTRICTED, PRIVATE]")
+                    + "`uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.SecurityClassification` "
+                    + "from String \"someInvalidEnumValue\": not one of the values accepted for Enum class: "
+                    + "[PUBLIC, RESTRICTED, PRIVATE]")
             .hasCauseInstanceOf(InvalidFormatException.class);
 
     }
@@ -850,7 +761,7 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes = getDefaultAttributesWithWorkType(formattedCreatedDate, formattedDueDate);
+        Map<String, Object> attributes = getDefaultAttributesWithWorkType(formattedCreatedDate, formattedDueDate);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
         Set<PermissionTypes> permissionsUnion = new HashSet<>(
@@ -860,17 +771,7 @@ class CFTTaskMapperTest {
                 PermissionTypes.MANAGE,
                 PermissionTypes.EXECUTE,
                 PermissionTypes.CANCEL,
-                PermissionTypes.REFER,
-                PermissionTypes.CLAIM,
-                PermissionTypes.ASSIGN,
-                PermissionTypes.UNASSIGN,
-                PermissionTypes.UNASSIGN_ASSIGN,
-                PermissionTypes.COMPLETE,
-                PermissionTypes.COMPLETE_OWN,
-                PermissionTypes.CANCEL_OWN,
-                PermissionTypes.UNASSIGN_CLAIM,
-                PermissionTypes.UNCLAIM,
-                PermissionTypes.UNCLAIM_ASSIGN
+                PermissionTypes.REFER
             )
         );
         Task task = cftTaskMapper.mapToTaskWithPermissions(taskResource, permissionsUnion);
@@ -889,7 +790,7 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate, null);
+        Map<String, Object> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate, null);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
         Set<PermissionTypes> permissionsUnion = new HashSet<>(
@@ -917,7 +818,7 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes =
+        Map<String, Object> attributes =
             getDefaultAttributesWithoutWithWorkType(formattedCreatedDate, formattedDueDate);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
@@ -949,19 +850,21 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes =
+        Map<String, Object> attributes =
             getDefaultAttributesWithoutWithWorkType(formattedCreatedDate, formattedDueDate);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
         taskResource.setSecurityClassification(null);
 
         Set<PermissionTypes> permissionsUnion = new HashSet<>(
-            asList(PermissionTypes.READ,
+            asList(
+                PermissionTypes.READ,
                 PermissionTypes.OWN,
                 PermissionTypes.MANAGE,
                 PermissionTypes.EXECUTE,
                 PermissionTypes.CANCEL,
-                PermissionTypes.REFER)
+                PermissionTypes.REFER
+            )
         );
         Task task = cftTaskMapper.mapToTaskWithPermissions(taskResource, permissionsUnion);
 
@@ -980,19 +883,21 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes =
+        Map<String, Object> attributes =
             getDefaultAttributesWithoutWithWorkType(formattedCreatedDate, formattedDueDate);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
         taskResource.setTaskSystem(null);
 
         Set<PermissionTypes> permissionsUnion = new HashSet<>(
-            asList(PermissionTypes.READ,
+            asList(
+                PermissionTypes.READ,
                 PermissionTypes.OWN,
                 PermissionTypes.MANAGE,
                 PermissionTypes.EXECUTE,
                 PermissionTypes.CANCEL,
-                PermissionTypes.REFER)
+                PermissionTypes.REFER
+            )
         );
         Task task = cftTaskMapper.mapToTaskWithPermissions(taskResource, permissionsUnion);
         assertNull(task.getTaskSystem());
@@ -1008,19 +913,21 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes =
+        Map<String, Object> attributes =
             getDefaultAttributesWithoutWithWorkType(formattedCreatedDate, formattedDueDate);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
         taskResource.setExecutionTypeCode(null);
 
         Set<PermissionTypes> permissionsUnion = new HashSet<>(
-            asList(PermissionTypes.READ,
+            asList(
+                PermissionTypes.READ,
                 PermissionTypes.OWN,
                 PermissionTypes.MANAGE,
                 PermissionTypes.EXECUTE,
                 PermissionTypes.CANCEL,
-                PermissionTypes.REFER)
+                PermissionTypes.REFER
+            )
         );
         Task task = cftTaskMapper.mapToTaskWithPermissions(taskResource, permissionsUnion);
         assertNull(task.getExecutionType());
@@ -1031,12 +938,12 @@ class CFTTaskMapperTest {
 
     @Test
     void should_throw_exception_when_map_task_resource_to_task_and_due_date_is_null() {
-        List<TaskAttribute> attributes = getDefaultAttributesWithoutDueDate();
+        Map<String, Object> attributes = getDefaultAttributesWithoutDueDate();
 
         AssertionsForClassTypes.assertThatThrownBy(() -> cftTaskMapper.mapToTaskResource(taskId, attributes))
             .isInstanceOf(NullPointerException.class)
             .hasNoCause()
-            .hasMessage("TASK_DUE_DATE must not be null");
+            .hasMessage("DUE_DATE must not be null");
     }
 
     @Test
@@ -1045,7 +952,7 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = ZonedDateTime.now().plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate, null);
+        Map<String, Object> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate, null);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
         assertNotNull(taskResource.getCreated());
@@ -1071,7 +978,7 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate, null);
+        Map<String, Object> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate, null);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
         assertNotNull(taskResource.getDueDateTime());
@@ -1242,7 +1149,7 @@ class CFTTaskMapperTest {
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
 
-        List<TaskAttribute> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate, null);
+        Map<String, Object> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate, null);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
         Set<PermissionTypes> permissionsUnion = new HashSet<>(
@@ -1296,8 +1203,9 @@ class CFTTaskMapperTest {
         ZonedDateTime priorityDate = createdDate.plusDays(3);
         String formattedPriorityDate = CAMUNDA_DATA_TIME_FORMATTER.format(priorityDate);
 
-        List<TaskAttribute> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate,
-            formattedPriorityDate);
+        Map<String, Object> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate,
+                                                              formattedPriorityDate
+        );
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
         Set<PermissionTypes> permissionsUnion = new HashSet<>(
@@ -1324,8 +1232,9 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate,
-                                                              null);
+        Map<String, Object> attributes = getDefaultAttributes(formattedCreatedDate, formattedDueDate,
+                                                              null
+        );
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
         taskResource.setPriorityDate(null);
@@ -1353,7 +1262,7 @@ class CFTTaskMapperTest {
         ZonedDateTime dueDate = createdDate.plusDays(1);
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
 
-        List<TaskAttribute> attributes = getDefaultAttributesWithWorkType(formattedCreatedDate, formattedDueDate);
+        Map<String, Object> attributes = getDefaultAttributesWithWorkType(formattedCreatedDate, formattedDueDate);
 
         TaskResource taskResource = cftTaskMapper.mapToTaskResource(taskId, attributes);
         Map<String, Object> taskAttributes = cftTaskMapper.getTaskAttributes(taskResource);
@@ -1392,114 +1301,6 @@ class CFTTaskMapperTest {
     }
 
     @Test
-    void should_extract_permission_union_all_true_with_non_granular_permission() {
-
-        TaskRoleResource taskRoleResource = new TaskRoleResource(
-            "tribunal-caseworker",
-            true,
-            true,
-            true,
-            true,
-            true,
-            true,
-            new String[]{},
-            0,
-            false,
-            "JUDICIAL",
-            taskId,
-            OffsetDateTime.parse("2021-05-09T20:15:45.345875+01:00"),
-            true,
-            false,
-            false,
-            true,
-            false,
-            true,
-            false,
-            false,
-            false,
-            true
-        );
-        Set<TaskRoleResource> taskRoleResources = new HashSet<>(singletonList(taskRoleResource));
-        List<RoleAssignment> roleAssignments = singletonList(RoleAssignmentCreator.aRoleAssignment().build());
-
-        Set<PermissionTypes> permissionsUnion =
-            cftTaskMapper.extractUnionOfPermissionsForUser(taskRoleResources, roleAssignments, false);
-
-        assertFalse(permissionsUnion.isEmpty());
-        assertTrue(permissionsUnion.contains(PermissionTypes.READ));
-        assertTrue(permissionsUnion.contains(PermissionTypes.OWN));
-        assertTrue(permissionsUnion.contains(PermissionTypes.MANAGE));
-        assertTrue(permissionsUnion.contains(PermissionTypes.EXECUTE));
-        assertTrue(permissionsUnion.contains(PermissionTypes.CANCEL));
-        assertTrue(permissionsUnion.contains(PermissionTypes.REFER));
-
-        assertFalse(permissionsUnion.contains(PermissionTypes.COMPLETE));
-        assertFalse(permissionsUnion.contains(PermissionTypes.COMPLETE_OWN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.CANCEL_OWN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.CLAIM));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNCLAIM));
-        assertFalse(permissionsUnion.contains(PermissionTypes.ASSIGN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNASSIGN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNCLAIM_ASSIGN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNASSIGN_CLAIM));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNASSIGN_ASSIGN));
-    }
-
-    @Test
-    void should_extract_permission_union_all_true_with_granular_permission() {
-
-        TaskRoleResource taskRoleResource = new TaskRoleResource(
-            "tribunal-caseworker",
-            true,
-            true,
-            true,
-            true,
-            true,
-            true,
-            new String[]{},
-            0,
-            false,
-            "JUDICIAL",
-            taskId,
-            OffsetDateTime.parse("2021-05-09T20:15:45.345875+01:00"),
-            true,
-            false,
-            false,
-            true,
-            false,
-            true,
-            false,
-            false,
-            false,
-            true
-        );
-        Set<TaskRoleResource> taskRoleResources = new HashSet<>(singletonList(taskRoleResource));
-        List<RoleAssignment> roleAssignments = singletonList(RoleAssignmentCreator.aRoleAssignment().build());
-
-        Set<PermissionTypes> permissionsUnion =
-            cftTaskMapper.extractUnionOfPermissionsForUser(taskRoleResources, roleAssignments, true);
-
-        assertFalse(permissionsUnion.isEmpty());
-        assertTrue(permissionsUnion.contains(PermissionTypes.READ));
-        assertTrue(permissionsUnion.contains(PermissionTypes.OWN));
-        assertTrue(permissionsUnion.contains(PermissionTypes.MANAGE));
-        assertTrue(permissionsUnion.contains(PermissionTypes.EXECUTE));
-        assertTrue(permissionsUnion.contains(PermissionTypes.CANCEL));
-        assertTrue(permissionsUnion.contains(PermissionTypes.REFER));
-
-        assertTrue(permissionsUnion.contains(PermissionTypes.COMPLETE));
-        assertFalse(permissionsUnion.contains(PermissionTypes.COMPLETE_OWN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.CANCEL_OWN));
-        assertTrue(permissionsUnion.contains(PermissionTypes.CLAIM));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNCLAIM));
-        assertTrue(permissionsUnion.contains(PermissionTypes.ASSIGN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNASSIGN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNCLAIM_ASSIGN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNASSIGN_CLAIM));
-        assertTrue(permissionsUnion.contains(PermissionTypes.UNASSIGN_ASSIGN));
-    }
-
-    @Test
     void should_extract_union_and_map_to_task() {
 
 
@@ -1513,19 +1314,6 @@ class CFTTaskMapperTest {
             false,
             new String[]{},
             0,
-            false,
-            "JUDICIAL",
-            taskId,
-            OffsetDateTime.parse("2021-05-09T20:15:45.345875+01:00"),
-            true,
-            false,
-            false,
-            true,
-            false,
-            true,
-            false,
-            false,
-            false,
             true
         );
         TaskRoleResource taskRoleResource2 = new TaskRoleResource(
@@ -1538,142 +1326,6 @@ class CFTTaskMapperTest {
             true,
             new String[]{},
             0,
-            true,
-            "JUDICIAL",
-            taskId,
-            OffsetDateTime.parse("2021-05-09T20:15:45.345875+01:00"),
-            false,
-            true,
-            true,
-            false,
-            false,
-            false,
-            false,
-            false,
-            false,
-            true
-        );
-
-        Set<TaskRoleResource> taskRoleResources = new HashSet<>(asList(taskRoleResource1, taskRoleResource2));
-        TaskResource taskResource = new TaskResource(
-            "taskId",
-            "aTaskName",
-            "startAppeal",
-            OffsetDateTime.parse("2022-05-09T20:15:45.345875+01:00"),
-            CFTTaskState.COMPLETED,
-            TaskSystem.SELF,
-            SecurityClassification.PUBLIC,
-            "title",
-            "a description",
-            null,
-            0,
-            0,
-            "someAssignee",
-            false,
-            new ExecutionTypeResource(ExecutionType.MANUAL, "Manual", "Manual Description"),
-            new WorkTypeResource("routine_work", "Routine work"),
-            "JUDICIAL",
-            false,
-            OffsetDateTime.parse("2022-05-09T20:15:45.345875+01:00"),
-            "1623278362430412",
-            "Asylum",
-            "TestCase",
-            "IA",
-            "1",
-            "TestRegion",
-            "765324",
-            "Taylor House",
-            BusinessContext.CFT_TASK,
-            "Some termination reason",
-            OffsetDateTime.parse("2021-05-09T20:15:45.345875+01:00"),
-            taskRoleResources,
-            "caseCategory",
-            EXPECTED_ADDITIONAL_PROPERTIES,
-                "nextHearingId",
-            OffsetDateTime.parse("2021-05-09T20:15:45.345875+01:00"),
-            OffsetDateTime.parse("2021-05-09T20:15:45.345875+01:00")
-        );
-
-        List<RoleAssignment> roleAssignments = singletonList(RoleAssignmentCreator.aRoleAssignment().build());
-
-        Task mappedTask = cftTaskMapper.mapToTaskAndExtractPermissionsUnion(taskResource, roleAssignments, false);
-
-        assertNotNull(mappedTask);
-
-        TaskPermissions taskPermissions = mappedTask.getPermissions();
-        assertNotNull(taskPermissions.getValues());
-
-        Set<PermissionTypes> permissionsUnion = taskPermissions.getValues();
-        assertTrue(permissionsUnion.contains(PermissionTypes.READ));
-        assertTrue(permissionsUnion.contains(PermissionTypes.OWN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.EXECUTE));
-        assertFalse(permissionsUnion.contains(PermissionTypes.MANAGE));
-        assertFalse(permissionsUnion.contains(PermissionTypes.CANCEL));
-        assertFalse(permissionsUnion.contains(PermissionTypes.REFER));
-
-        assertFalse(permissionsUnion.contains(PermissionTypes.COMPLETE));
-        assertFalse(permissionsUnion.contains(PermissionTypes.COMPLETE_OWN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.CANCEL_OWN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.CLAIM));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNCLAIM));
-        assertFalse(permissionsUnion.contains(PermissionTypes.ASSIGN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNASSIGN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNCLAIM_ASSIGN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNASSIGN_CLAIM));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNASSIGN_ASSIGN));
-    }
-
-    @Test
-    void should_extract_union_and_map_to_task_granular_permission() {
-
-        TaskRoleResource taskRoleResource1 = new TaskRoleResource(
-            "tribunal-caseworker",
-            true,
-            true,
-            false,
-            false,
-            false,
-            false,
-            new String[]{"SPECIFIC", "STANDARD"},
-            0,
-            false,
-            "JUDICIAL",
-            taskId,
-            OffsetDateTime.parse("2021-05-09T20:15:45.345875+01:00"),
-            true,
-            false,
-            false,
-            true,
-            false,
-            true,
-            false,
-            false,
-            false,
-            true
-        );
-        TaskRoleResource taskRoleResource2 = new TaskRoleResource(
-            "senior-tribunal-caseworker",
-            false,
-            true,
-            false,
-            true,
-            true,
-            true,
-            new String[]{},
-            0,
-            true,
-            "JUDICIAL",
-            taskId,
-            OffsetDateTime.parse("2021-05-09T20:15:45.345875+01:00"),
-            false,
-            true,
-            true,
-            false,
-            false,
-            false,
-            false,
-            false,
-            false,
             true
         );
 
@@ -1719,7 +1371,7 @@ class CFTTaskMapperTest {
 
         List<RoleAssignment> roleAssignments = singletonList(RoleAssignmentCreator.aRoleAssignment().build());
 
-        Task mappedTask = cftTaskMapper.mapToTaskAndExtractPermissionsUnion(taskResource, roleAssignments, true);
+        Task mappedTask = cftTaskMapper.mapToTaskAndExtractPermissionsUnion(taskResource, roleAssignments,false);
 
         assertNotNull(mappedTask);
 
@@ -1733,17 +1385,6 @@ class CFTTaskMapperTest {
         assertFalse(permissionsUnion.contains(PermissionTypes.MANAGE));
         assertFalse(permissionsUnion.contains(PermissionTypes.CANCEL));
         assertFalse(permissionsUnion.contains(PermissionTypes.REFER));
-
-        assertTrue(permissionsUnion.contains(PermissionTypes.COMPLETE));
-        assertFalse(permissionsUnion.contains(PermissionTypes.COMPLETE_OWN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.CANCEL_OWN));
-        assertTrue(permissionsUnion.contains(PermissionTypes.CLAIM));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNCLAIM));
-        assertTrue(permissionsUnion.contains(PermissionTypes.ASSIGN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNASSIGN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNCLAIM_ASSIGN));
-        assertFalse(permissionsUnion.contains(PermissionTypes.UNASSIGN_CLAIM));
-        assertTrue(permissionsUnion.contains(PermissionTypes.UNASSIGN_ASSIGN));
     }
 
     @Test
@@ -2319,62 +1960,6 @@ class CFTTaskMapperTest {
         assertEquals("nextHearingId", taskResource.getNextHearingId());
     }
 
-    @Test
-    void should_map_task_role_permissions_for_granular_permissions() {
-        TaskRoleResource roleResource = new TaskRoleResource(
-            "ctsc",
-            true,
-            true,
-            true,
-            true,
-            true,
-            true,
-            new String[]{"SPECIFIC", "STANDARD"},
-            0,
-            false,
-            "CTSC",
-            "taskId",
-            OffsetDateTime.parse("2021-05-09T20:15:45.345875+01:00"),
-            true,
-            true,
-            true,
-            true,
-            true,
-            true,
-            true,
-            true,
-            true,
-            true
-        );
-
-        final TaskRolePermissions taskRolePermissions = cftTaskMapper.mapToTaskRolePermissions(roleResource, true);
-
-        assertEquals("ctsc", taskRolePermissions.getRoleName());
-        assertEquals("CTSC", taskRolePermissions.getRoleCategory());
-        assertNotNull(taskRolePermissions.getPermissions());
-        assertFalse(taskRolePermissions.getPermissions().isEmpty());
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.READ));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.OWN));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.MANAGE));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.EXECUTE));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.CANCEL));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.REFER));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.CLAIM));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.ASSIGN));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.UNASSIGN));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.UNASSIGN_ASSIGN));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.COMPLETE));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.COMPLETE_OWN));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.CANCEL_OWN));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.UNASSIGN_CLAIM));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.UNCLAIM));
-        assertTrue(taskRolePermissions.getPermissions().contains(PermissionTypes.UNCLAIM_ASSIGN));
-
-        assertFalse(taskRolePermissions.getAuthorisations().isEmpty());
-        assertTrue(taskRolePermissions.getAuthorisations().contains("SPECIFIC"));
-        assertTrue(taskRolePermissions.getAuthorisations().contains("STANDARD"));
-    }
-
     private TaskResource createTaskResource() {
         return new TaskResource(
             "taskId",
@@ -2545,255 +2130,182 @@ class CFTTaskMapperTest {
             singleton(roleResource),
             "caseCategory",
             EXPECTED_ADDITIONAL_PROPERTIES,
-                "nextHearingId",
+            "nextHearingId",
             OffsetDateTime.parse("2021-05-09T20:15:45.345875+01:00"),
             OffsetDateTime.parse("2021-05-09T20:15:45.345875+01:00")
         );
     }
 
-    private List<TaskAttribute> getDefaultAttributes(String createdDate, String dueDate, String priorityDate) {
-        return asList(
-            new TaskAttribute(TaskAttributeDefinition.TASK_ASSIGNEE, "someAssignee"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_AUTO_ASSIGNED, false),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_CATEGORY, "someCaseCategory"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_ID, "00000"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_NAME, "someCaseName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_TYPE_ID, "someCaseType"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CREATED, createdDate),
-            new TaskAttribute(TaskAttributeDefinition.TASK_DUE_DATE, dueDate),
-            new TaskAttribute(TaskAttributeDefinition.TASK_DESCRIPTION, "someCamundaTaskDescription"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_EXECUTION_TYPE_NAME, "MANUAL"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_HAS_WARNINGS, false),
-            new TaskAttribute(TaskAttributeDefinition.TASK_JURISDICTION, "someJurisdiction"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_LOCATION, "someStaffLocationId"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_LOCATION_NAME, "someStaffLocationName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NAME, "someCamundaTaskName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_REGION, "someRegion"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SECURITY_CLASSIFICATION, "PUBLIC"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_STATE, CFTTaskState.UNCONFIGURED),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SYSTEM, "SELF"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TITLE, "someTitle"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TYPE, "someTaskType"),
-            //Unmapped
-            new TaskAttribute(TaskAttributeDefinition.TASK_ASSIGNMENT_EXPIRY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_BUSINESS_CONTEXT, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_PRIORITY_DATE, priorityDate),
-            new TaskAttribute(TaskAttributeDefinition.TASK_MAJOR_PRIORITY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_MINOR_PRIORITY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ROLES, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ROLE_CATEGORY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_REGION_NAME, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TERMINATION_REASON, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_WORK_TYPE, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NOTES, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ADDITIONAL_PROPERTIES, EXPECTED_ADDITIONAL_PROPERTIES),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NEXT_HEARING_ID, "nextHearingId"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NEXT_HEARING_DATE,
-                              CAMUNDA_DATA_TIME_FORMATTER.format(ZonedDateTime.now()))
-        );
-    }
-
-    private Map<String, Object> getDefaultInitiationAttributes(String createdDate,
-                                                               String dueDate, String priorityDate) {
-
-        Map<String, Object> attributes = new HashMap<>();
-
-        attributes.put(ASSIGNEE.value(), "someAssignee");
-        attributes.put(AUTO_ASSIGNED.value(), false);
-        attributes.put(CASE_CATEGORY.value(), "someCaseCategory");
-        attributes.put(CASE_ID.value(), "00000");
-        attributes.put(CASE_NAME.value(), "someCaseName");
-        attributes.put(CASE_TYPE_ID.value(), "someCaseType");
-        attributes.put(CREATED.value(), createdDate);
-        attributes.put(DUE_DATE.value(), dueDate);
-        attributes.put(DESCRIPTION.value(), "someCamundaTaskDescription");
-        attributes.put(EXECUTION_TYPE.value(), "MANUAL");
-        attributes.put(HAS_WARNINGS.value(), true);
-        attributes.put(WARNING_LIST.value(), "[]");
-        attributes.put(JURISDICTION.value(), "someJurisdiction");
-        attributes.put(LOCATION.value(), "someStaffLocationId");
-        attributes.put(LOCATION_NAME.value(), "someStaffLocationName");
+    private Map<String, Object> getDefaultAttributes(String createdDate, String dueDate, String priorityDate) {
+        HashMap<String, Object> attributes = new HashMap<>();
+        attributes.put(CamundaVariableDefinition.ASSIGNEE.value(), "someAssignee");
+        attributes.put(CamundaVariableDefinition.AUTO_ASSIGNED.value(), false);
+        attributes.put(CamundaVariableDefinition.CASE_CATEGORY.value(), "someCaseCategory");
+        attributes.put(CamundaVariableDefinition.CASE_ID.value(), "00000");
+        attributes.put(CamundaVariableDefinition.CASE_NAME.value(), "someCaseName");
+        attributes.put(CamundaVariableDefinition.CASE_TYPE_ID.value(), "someCaseType");
+        if (createdDate != null) {
+            attributes.put(CamundaVariableDefinition.CREATED
+                               .value(), createdDate);
+        }
+        if (dueDate != null) {
+            attributes.put(CamundaVariableDefinition.DUE_DATE.value(), dueDate);
+        }
+        if (priorityDate != null) {
+            attributes.put(CamundaVariableDefinition.PRIORITY_DATE.value(), priorityDate);
+        }
+        attributes.put(CamundaVariableDefinition.DESCRIPTION.value(), "someCamundaTaskDescription");
+        attributes.put(CamundaVariableDefinition.EXECUTION_TYPE.value(), "MANUAL");
+        attributes.put(CamundaVariableDefinition.HAS_WARNINGS.value(), false);
+        attributes.put(CamundaVariableDefinition.JURISDICTION.value(), "someJurisdiction");
+        attributes.put(CamundaVariableDefinition.LOCATION.value(), "someStaffLocationId");
+        attributes.put(CamundaVariableDefinition.LOCATION_NAME.value(), "someStaffLocationName");
         attributes.put(CamundaVariableDefinition.TASK_NAME.value(), "someCamundaTaskName");
-        attributes.put(REGION.value(), "someRegion");
-        attributes.put(SECURITY_CLASSIFICATION.value(), "PUBLIC");
-        attributes.put(TASK_STATE.value(), CFTTaskState.UNCONFIGURED);
-        attributes.put(TASK_SYSTEM.value(), "SELF");
-        attributes.put(TITLE.value(), "someTitle");
-        attributes.put(TASK_TYPE.value(), "someTaskType");
-        attributes.put(ADDITIONAL_PROPERTIES.value(), EXPECTED_ADDITIONAL_PROPERTIES);
-        attributes.put(PRIORITY_DATE.value(), priorityDate);
-        attributes.put(NEXT_HEARING_ID.value(), "nextHearingId");
-        attributes.put(NEXT_HEARING_DATE.value(), CAMUNDA_DATA_TIME_FORMATTER.format(ZonedDateTime.now()));
-
+        attributes.put(CamundaVariableDefinition.REGION.value(), "someRegion");
+        attributes.put(CamundaVariableDefinition.SECURITY_CLASSIFICATION.value(), "PUBLIC");
+        attributes.put(CamundaVariableDefinition.TASK_STATE.value(), CFTTaskState.UNCONFIGURED);
+        attributes.put(CamundaVariableDefinition.TASK_SYSTEM.value(), "SELF");
+        attributes.put(CamundaVariableDefinition.TITLE.value(), "someTitle");
+        attributes.put(CamundaVariableDefinition.TASK_TYPE.value(), "someTaskType");
+        attributes.put(CamundaVariableDefinition.ADDITIONAL_PROPERTIES.value(), EXPECTED_ADDITIONAL_PROPERTIES);
+        attributes.put(CamundaVariableDefinition.NEXT_HEARING_ID.value(), "nextHearingId");
+        attributes.put(
+            CamundaVariableDefinition.NEXT_HEARING_DATE.value(),
+            CAMUNDA_DATA_TIME_FORMATTER.format(ZonedDateTime.now())
+        );
         return attributes;
     }
 
-
-    private List<TaskAttribute> getDefaultAttributesWithoutDueDate() {
-        return asList(
-            new TaskAttribute(TaskAttributeDefinition.TASK_ASSIGNEE, "someAssignee"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_AUTO_ASSIGNED, false),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_CATEGORY, "someCaseCategory"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_ID, "00000"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_NAME, "someCaseName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_TYPE_ID, "someCaseType"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CREATED, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_DUE_DATE, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_DESCRIPTION, "someCamundaTaskDescription"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_EXECUTION_TYPE_NAME, "MANUAL"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_HAS_WARNINGS, false),
-            new TaskAttribute(TaskAttributeDefinition.TASK_JURISDICTION, "someJurisdiction"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_LOCATION, "someStaffLocationId"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_LOCATION_NAME, "someStaffLocationName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NAME, "someCamundaTaskName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_REGION, "someRegion"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SECURITY_CLASSIFICATION, "PUBLIC"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_STATE, CFTTaskState.UNCONFIGURED),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SYSTEM, "SELF"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TITLE, "someTitle"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TYPE, "someTaskType"),
-            //Unmapped
-            new TaskAttribute(TaskAttributeDefinition.TASK_ASSIGNMENT_EXPIRY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_BUSINESS_CONTEXT, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_MAJOR_PRIORITY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_MINOR_PRIORITY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ROLES, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ROLE_CATEGORY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_REGION_NAME, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TERMINATION_REASON, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_WORK_TYPE, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NOTES, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ADDITIONAL_PROPERTIES, EXPECTED_ADDITIONAL_PROPERTIES),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NEXT_HEARING_ID, "nextHearingId"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NEXT_HEARING_DATE,
-                              CAMUNDA_DATA_TIME_FORMATTER.format(ZonedDateTime.now()))
+    private Map<String, Object> getDefaultAttributesWithoutDueDate() {
+        Map<String, Object> attributes = new HashMap<>();
+        attributes.put(CamundaVariableDefinition.ASSIGNEE.value(), "someAssignee");
+        attributes.put(CamundaVariableDefinition.AUTO_ASSIGNED.value(), false);
+        attributes.put(CamundaVariableDefinition.CASE_CATEGORY.value(), "someCaseCategory");
+        attributes.put(CamundaVariableDefinition.CASE_ID.value(), "00000");
+        attributes.put(CamundaVariableDefinition.CASE_NAME.value(), "someCaseName");
+        attributes.put(CamundaVariableDefinition.CASE_TYPE_ID.value(), "someCaseType");
+        attributes.put(CamundaVariableDefinition.DESCRIPTION.value(), "someCamundaTaskDescription");
+        attributes.put(CamundaVariableDefinition.EXECUTION_TYPE.value(), "MANUAL");
+        attributes.put(CamundaVariableDefinition.HAS_WARNINGS.value(), false);
+        attributes.put(CamundaVariableDefinition.JURISDICTION.value(), "someJurisdiction");
+        attributes.put(CamundaVariableDefinition.LOCATION.value(), "someStaffLocationId");
+        attributes.put(CamundaVariableDefinition.LOCATION_NAME.value(), "someStaffLocationName");
+        attributes.put(CamundaVariableDefinition.TASK_NAME.value(), "someCamundaTaskName");
+        attributes.put(CamundaVariableDefinition.REGION.value(), "someRegion");
+        attributes.put(CamundaVariableDefinition.SECURITY_CLASSIFICATION.value(), "PUBLIC");
+        attributes.put(CamundaVariableDefinition.TASK_STATE.value(), CFTTaskState.UNCONFIGURED);
+        attributes.put(CamundaVariableDefinition.TASK_SYSTEM.value(), "SELF");
+        attributes.put(CamundaVariableDefinition.TITLE.value(), "someTitle");
+        attributes.put(CamundaVariableDefinition.TASK_TYPE.value(), "someTaskType");
+        attributes.put(CamundaVariableDefinition.ADDITIONAL_PROPERTIES.value(), EXPECTED_ADDITIONAL_PROPERTIES);
+        attributes.put(CamundaVariableDefinition.NEXT_HEARING_ID.value(), "nextHearingId");
+        attributes.put(
+            CamundaVariableDefinition.NEXT_HEARING_DATE.value(),
+            CAMUNDA_DATA_TIME_FORMATTER.format(ZonedDateTime.now())
         );
+        return attributes;
     }
 
-    private List<TaskAttribute> getDefaultAttributesWithWorkType(String createdDate, String dueDate) {
-        return asList(
-            new TaskAttribute(TaskAttributeDefinition.TASK_ASSIGNEE, "someAssignee"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_AUTO_ASSIGNED, false),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_CATEGORY, "someCaseCategory"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_ID, "00000"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_NAME, "someCaseName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_TYPE_ID, "someCaseType"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CREATED, createdDate),
-            new TaskAttribute(TaskAttributeDefinition.TASK_DUE_DATE, dueDate),
-            new TaskAttribute(TaskAttributeDefinition.TASK_DESCRIPTION, "someCamundaTaskDescription"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_EXECUTION_TYPE_NAME, "MANUAL"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_HAS_WARNINGS, false),
-            new TaskAttribute(TaskAttributeDefinition.TASK_JURISDICTION, "someJurisdiction"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_LOCATION, "someStaffLocationId"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_LOCATION_NAME, "someStaffLocationName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NAME, "someCamundaTaskName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_REGION, "someRegion"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SECURITY_CLASSIFICATION, "PUBLIC"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_STATE, CFTTaskState.UNCONFIGURED),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SYSTEM, "SELF"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TITLE, "someTitle"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TYPE, "someTaskType"),
-            //Unmapped
-            new TaskAttribute(TaskAttributeDefinition.TASK_ASSIGNMENT_EXPIRY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_BUSINESS_CONTEXT, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_MAJOR_PRIORITY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_MINOR_PRIORITY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ROLES, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ROLE_CATEGORY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_REGION_NAME, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TERMINATION_REASON, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_WORK_TYPE, "someWorkType"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NOTES, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ADDITIONAL_PROPERTIES, EXPECTED_ADDITIONAL_PROPERTIES),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NEXT_HEARING_ID, "nextHearingId"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NEXT_HEARING_DATE,
-                              CAMUNDA_DATA_TIME_FORMATTER.format(ZonedDateTime.now()))
+    private Map<String, Object> getDefaultAttributesWithWorkType(String createdDate, String dueDate) {
+        Map<String, Object> attributes = new HashMap<>();
+        attributes.put(CamundaVariableDefinition.ASSIGNEE.value(), "someAssignee");
+        attributes.put(CamundaVariableDefinition.AUTO_ASSIGNED.value(), false);
+        attributes.put(CamundaVariableDefinition.CASE_CATEGORY.value(), "someCaseCategory");
+        attributes.put(CamundaVariableDefinition.CASE_ID.value(), "00000");
+        attributes.put(CamundaVariableDefinition.CASE_NAME.value(), "someCaseName");
+        attributes.put(CamundaVariableDefinition.CASE_TYPE_ID.value(), "someCaseType");
+        attributes.put(CamundaVariableDefinition.CREATED.value(), createdDate);
+        attributes.put(CamundaVariableDefinition.DUE_DATE.value(), dueDate);
+        attributes.put(CamundaVariableDefinition.DESCRIPTION.value(), "someCamundaTaskDescription");
+        attributes.put(CamundaVariableDefinition.EXECUTION_TYPE.value(), "MANUAL");
+        attributes.put(CamundaVariableDefinition.HAS_WARNINGS.value(), false);
+        attributes.put(CamundaVariableDefinition.JURISDICTION.value(), "someJurisdiction");
+        attributes.put(CamundaVariableDefinition.LOCATION.value(), "someStaffLocationId");
+        attributes.put(CamundaVariableDefinition.LOCATION_NAME.value(), "someStaffLocationName");
+        attributes.put(CamundaVariableDefinition.TASK_NAME.value(), "someCamundaTaskName");
+        attributes.put(CamundaVariableDefinition.REGION.value(), "someRegion");
+        attributes.put(CamundaVariableDefinition.SECURITY_CLASSIFICATION.value(), "PUBLIC");
+        attributes.put(CamundaVariableDefinition.TASK_STATE.value(), CFTTaskState.UNCONFIGURED);
+        attributes.put(CamundaVariableDefinition.TASK_SYSTEM.value(), "SELF");
+        attributes.put(CamundaVariableDefinition.TITLE.value(), "someTitle");
+        attributes.put(CamundaVariableDefinition.TASK_TYPE.value(), "someTaskType");
+        attributes.put(CamundaVariableDefinition.WORK_TYPE.value(), "someWorkType");
+        attributes.put(CamundaVariableDefinition.ADDITIONAL_PROPERTIES.value(), EXPECTED_ADDITIONAL_PROPERTIES);
+        attributes.put(CamundaVariableDefinition.NEXT_HEARING_ID.value(), "nextHearingId");
+        attributes.put(
+            CamundaVariableDefinition.NEXT_HEARING_DATE.value(),
+            CAMUNDA_DATA_TIME_FORMATTER.format(ZonedDateTime.now())
         );
+        return attributes;
     }
 
-    private List<TaskAttribute> getDefaultAttributesWithoutWithWorkType(String createdDate, String dueDate) {
-        return asList(
-            new TaskAttribute(TaskAttributeDefinition.TASK_ASSIGNEE, "someAssignee"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_AUTO_ASSIGNED, false),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_CATEGORY, "someCaseCategory"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_ID, "00000"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_NAME, "someCaseName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_TYPE_ID, "someCaseType"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CREATED, createdDate),
-            new TaskAttribute(TaskAttributeDefinition.TASK_DUE_DATE, dueDate),
-            new TaskAttribute(TaskAttributeDefinition.TASK_DESCRIPTION, "someCamundaTaskDescription"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_EXECUTION_TYPE_NAME, "MANUAL"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_HAS_WARNINGS, false),
-            new TaskAttribute(TaskAttributeDefinition.TASK_JURISDICTION, "someJurisdiction"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_LOCATION, "someStaffLocationId"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_LOCATION_NAME, "someStaffLocationName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NAME, "someCamundaTaskName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_REGION, "someRegion"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SECURITY_CLASSIFICATION, "PUBLIC"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_STATE, CFTTaskState.UNCONFIGURED),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SYSTEM, "SELF"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TITLE, "someTitle"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TYPE, "someTaskType"),
-            //Unmapped
-            new TaskAttribute(TaskAttributeDefinition.TASK_ASSIGNMENT_EXPIRY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_BUSINESS_CONTEXT, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_MAJOR_PRIORITY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_MINOR_PRIORITY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ROLES, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ROLE_CATEGORY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_REGION_NAME, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TERMINATION_REASON, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NOTES, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ADDITIONAL_PROPERTIES, EXPECTED_ADDITIONAL_PROPERTIES),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NEXT_HEARING_ID, "nextHearingId"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NEXT_HEARING_DATE,
-                              CAMUNDA_DATA_TIME_FORMATTER.format(ZonedDateTime.now()))
+    private Map<String, Object> getDefaultAttributesWithoutWithWorkType(String createdDate, String dueDate) {
+        Map<String, Object> attributes = new HashMap<>();
+        attributes.put(CamundaVariableDefinition.ASSIGNEE.value(), "someAssignee");
+        attributes.put(CamundaVariableDefinition.AUTO_ASSIGNED.value(), false);
+        attributes.put(CamundaVariableDefinition.CASE_CATEGORY.value(), "someCaseCategory");
+        attributes.put(CamundaVariableDefinition.CASE_ID.value(), "00000");
+        attributes.put(CamundaVariableDefinition.CASE_NAME.value(), "someCaseName");
+        attributes.put(CamundaVariableDefinition.CASE_TYPE_ID.value(), "someCaseType");
+        attributes.put(CamundaVariableDefinition.CREATED
+                           .value(), createdDate);
+        attributes.put(CamundaVariableDefinition.DUE_DATE.value(), dueDate);
+        attributes.put(CamundaVariableDefinition.DESCRIPTION.value(), "someCamundaTaskDescription");
+        attributes.put(CamundaVariableDefinition.EXECUTION_TYPE.value(), "MANUAL");
+        attributes.put(CamundaVariableDefinition.HAS_WARNINGS.value(), false);
+        attributes.put(CamundaVariableDefinition.JURISDICTION.value(), "someJurisdiction");
+        attributes.put(CamundaVariableDefinition.LOCATION.value(), "someStaffLocationId");
+        attributes.put(CamundaVariableDefinition.LOCATION_NAME.value(), "someStaffLocationName");
+        attributes.put(CamundaVariableDefinition.TASK_NAME.value(), "someCamundaTaskName");
+        attributes.put(CamundaVariableDefinition.REGION.value(), "someRegion");
+        attributes.put(CamundaVariableDefinition.SECURITY_CLASSIFICATION.value(), "PUBLIC");
+        attributes.put(CamundaVariableDefinition.TASK_STATE.value(), CFTTaskState.UNCONFIGURED);
+        attributes.put(CamundaVariableDefinition.TASK_SYSTEM.value(), "SELF");
+        attributes.put(CamundaVariableDefinition.TITLE.value(), "someTitle");
+        attributes.put(CamundaVariableDefinition.TASK_TYPE.value(), "someTaskType");
+        attributes.put(CamundaVariableDefinition.ADDITIONAL_PROPERTIES.value(), EXPECTED_ADDITIONAL_PROPERTIES);
+        attributes.put(CamundaVariableDefinition.NEXT_HEARING_ID.value(), "nextHearingId");
+        attributes.put(
+            CamundaVariableDefinition.NEXT_HEARING_DATE.value(),
+            CAMUNDA_DATA_TIME_FORMATTER.format(ZonedDateTime.now())
         );
+        return attributes;
     }
 
-    private List<TaskAttribute> getDefaultAttributesWithWarnings(String createdDate, String dueDate) {
+    private Map<String, Object> getDefaultAttributesWithWarnings(String createdDate, String dueDate) {
         String values = "[{\"warningCode\":\"Code1\", \"warningText\":\"Text1\"}, "
-                        + "{\"warningCode\":\"Code2\", \"warningText\":\"Text2\"}]";
-        return asList(
-            new TaskAttribute(TaskAttributeDefinition.TASK_ASSIGNEE, "someAssignee"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_AUTO_ASSIGNED, false),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_CATEGORY, "someCaseCategory"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_ID, "00000"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_NAME, "someCaseName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CASE_TYPE_ID, "someCaseType"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_CREATED, createdDate),
-            new TaskAttribute(TaskAttributeDefinition.TASK_DUE_DATE, dueDate),
-            new TaskAttribute(TaskAttributeDefinition.TASK_DESCRIPTION, "someCamundaTaskDescription"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_EXECUTION_TYPE_NAME, "MANUAL"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_HAS_WARNINGS, true),
-            new TaskAttribute(TaskAttributeDefinition.TASK_JURISDICTION, "someJurisdiction"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_LOCATION, "someStaffLocationId"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_LOCATION_NAME, "someStaffLocationName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NAME, "someCamundaTaskName"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_REGION, "someRegion"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SECURITY_CLASSIFICATION, "PUBLIC"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_STATE, CFTTaskState.UNCONFIGURED),
-            new TaskAttribute(TaskAttributeDefinition.TASK_SYSTEM, "SELF"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TITLE, "someTitle"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TYPE, "someTaskType"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_WARNINGS, values),
-            //Unmapped
-            new TaskAttribute(TaskAttributeDefinition.TASK_ASSIGNMENT_EXPIRY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_BUSINESS_CONTEXT, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_MAJOR_PRIORITY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_MINOR_PRIORITY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ROLES, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ROLE_CATEGORY, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_REGION_NAME, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_TERMINATION_REASON, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_WORK_TYPE, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NOTES, null),
-            new TaskAttribute(TaskAttributeDefinition.TASK_ADDITIONAL_PROPERTIES, EXPECTED_ADDITIONAL_PROPERTIES),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NEXT_HEARING_ID, "nextHearingId"),
-            new TaskAttribute(TaskAttributeDefinition.TASK_NEXT_HEARING_DATE,
-                              CAMUNDA_DATA_TIME_FORMATTER.format(ZonedDateTime.now()))
+            + "{\"warningCode\":\"Code2\", \"warningText\":\"Text2\"}]";
+        Map<String, Object> attributes = new HashMap<>();
+        attributes.put(CamundaVariableDefinition.ASSIGNEE.value(), "someAssignee");
+        attributes.put(CamundaVariableDefinition.AUTO_ASSIGNED.value(), false);
+        attributes.put(CamundaVariableDefinition.CASE_CATEGORY.value(), "someCaseCategory");
+        attributes.put(CamundaVariableDefinition.CASE_ID.value(), "00000");
+        attributes.put(CamundaVariableDefinition.CASE_NAME.value(), "someCaseName");
+        attributes.put(CamundaVariableDefinition.CASE_TYPE_ID.value(), "someCaseType");
+        attributes.put(CamundaVariableDefinition.CREATED
+                           .value(), createdDate);
+        attributes.put(CamundaVariableDefinition.DUE_DATE.value(), dueDate);
+        attributes.put(CamundaVariableDefinition.DESCRIPTION.value(), "someCamundaTaskDescription");
+        attributes.put(CamundaVariableDefinition.EXECUTION_TYPE.value(), "MANUAL");
+        attributes.put(CamundaVariableDefinition.HAS_WARNINGS.value(), true);
+        attributes.put(CamundaVariableDefinition.JURISDICTION.value(), "someJurisdiction");
+        attributes.put(CamundaVariableDefinition.LOCATION.value(), "someStaffLocationId");
+        attributes.put(CamundaVariableDefinition.LOCATION_NAME.value(), "someStaffLocationName");
+        attributes.put(CamundaVariableDefinition.TASK_NAME.value(), "someCamundaTaskName");
+        attributes.put(CamundaVariableDefinition.REGION.value(), "someRegion");
+        attributes.put(CamundaVariableDefinition.SECURITY_CLASSIFICATION.value(), "PUBLIC");
+        attributes.put(CamundaVariableDefinition.TASK_STATE.value(), CFTTaskState.UNCONFIGURED);
+        attributes.put(CamundaVariableDefinition.TASK_SYSTEM.value(), "SELF");
+        attributes.put(CamundaVariableDefinition.TITLE.value(), "someTitle");
+        attributes.put(CamundaVariableDefinition.TASK_TYPE.value(), "someTaskType");
+        attributes.put(WARNING_LIST.value(), values);
+        attributes.put(CamundaVariableDefinition.ADDITIONAL_PROPERTIES.value(), EXPECTED_ADDITIONAL_PROPERTIES);
+        attributes.put(CamundaVariableDefinition.NEXT_HEARING_ID.value(), "nextHearingId");
+        attributes.put(
+            CamundaVariableDefinition.NEXT_HEARING_DATE.value(),
+            CAMUNDA_DATA_TIME_FORMATTER.format(ZonedDateTime.now())
         );
-
+        return attributes;
     }
 
     private String writeValueAsString(Map<String, String> data) {
