+++ src/test/java/uk/gov/hmcts/reform/wataskmanagementapi/services/TaskManagementServiceTest.java
@@ -35,15 +35,12 @@ import uk.gov.hmcts.reform.wataskmanagementapi.cft.enums.CFTTaskState;
 import uk.gov.hmcts.reform.wataskmanagementapi.cft.query.CftQueryService;
 import uk.gov.hmcts.reform.wataskmanagementapi.clients.CamundaServiceApi;
 import uk.gov.hmcts.reform.wataskmanagementapi.config.LaunchDarklyFeatureFlagProvider;
-import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.InitiateTaskRequestAttributes;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.InitiateTaskRequestMap;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.NotesRequest;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.TaskOperationRequest;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.entities.MarkTaskToReconfigureTaskFilter;
-import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.entities.TaskAttribute;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.entities.TaskFilter;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.entities.TaskOperation;
-import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskFilterOperator;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskOperationName;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.options.CompletionOptions;
@@ -74,7 +71,6 @@ import uk.gov.hmcts.reform.wataskmanagementapi.taskconfiguration.services.TaskAu
 import java.io.Serializable;
 import java.sql.SQLException;
 import java.time.OffsetDateTime;
-import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -144,9 +140,7 @@ import static uk.gov.hmcts.reform.wataskmanagementapi.cft.enums.CFTTaskState.TER
 import static uk.gov.hmcts.reform.wataskmanagementapi.config.features.FeatureFlag.GRANULAR_PERMISSION_FEATURE;
 import static uk.gov.hmcts.reform.wataskmanagementapi.config.features.FeatureFlag.RELEASE_4_GRANULAR_PERMISSION_RESPONSE;
 import static uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.InitiateTaskOperation.INITIATION;
-import static uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition.TASK_DUE_DATE;
 import static uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition.TASK_NAME;
-import static uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition.TASK_ROLE_ASSIGNMENT_ID;
 import static uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition.TASK_TYPE;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaTime.CAMUNDA_DATA_TIME_FORMATTER;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.DUE_DATE;
@@ -2795,143 +2789,6 @@ class TaskManagementServiceTest extends CamundaHelpers {
                 verify(cftTaskDatabaseService, times(0)).saveTask(taskResource);
             }
         }
-    }
-
-    @Nested
-    @DisplayName("OldInitiateTask()")
-    class OldInitiateTask {
-        ZonedDateTime dueDate = ZonedDateTime.now();
-        String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
-        private final InitiateTaskRequestAttributes initiateTaskRequest = new InitiateTaskRequestAttributes(
-            INITIATION,
-            asList(
-                new TaskAttribute(TASK_TYPE, A_TASK_TYPE),
-                new TaskAttribute(TASK_NAME, A_TASK_NAME),
-                new TaskAttribute(TASK_DUE_DATE, formattedDueDate),
-                new TaskAttribute(TASK_ROLE_ASSIGNMENT_ID, SOME_ROLE_ASSIGNMENT_ID)
-            )
-        );
-        @Mock
-        private TaskResource taskResource;
-
-        @Mock
-        private Map<String, Object> taskAttributes;
-
-        @Test
-        void given_some_error_other_than_DataAccessException_when_requiring_lock_then_throw_500_error()
-            throws SQLException {
-            doThrow(new RuntimeException("some unexpected error"))
-                .when(cftTaskDatabaseService).insertAndLock(anyString(), any());
-
-            assertThatThrownBy(() -> taskManagementService.initiateTask(taskId, initiateTaskRequest))
-                .isInstanceOf(RuntimeException.class);
-        }
-
-        @Test
-        void given_some_error_when_initiateTaskProcess_then_throw_500_error() {
-            when(cftTaskMapper.readDate(any(), any(TaskAttributeDefinition.class), any())).thenCallRealMethod();
-            doThrow(new RuntimeException("some unexpected error"))
-                .when(cftTaskMapper).mapToTaskResource(anyString(), anyList());
-
-            assertThatThrownBy(() -> taskManagementService.initiateTask(taskId, initiateTaskRequest))
-                .isInstanceOf(GenericServerErrorException.class)
-                .hasMessage("Generic Server Error: The action could not be completed "
-                            + "because there was a problem when initiating the task.");
-        }
-
-        @Test
-        void given_DataAccessException_when_initiate_task_then_throw_503_error() throws SQLException {
-            String msg = "duplicate key value violates unique constraint \"tasks_pkey\"";
-            doThrow(new DataIntegrityViolationException(msg))
-                .when(cftTaskDatabaseService).insertAndLock(anyString(), any());
-
-            assertThatThrownBy(() -> taskManagementService.initiateTask(taskId, initiateTaskRequest))
-                .isInstanceOf(DatabaseConflictException.class)
-                .hasMessage("Database Conflict Error: "
-                            + "The action could not be completed because there was a conflict in the database.");
-        }
-
-        @Test
-        void given_initiateTask_task_is_initiated() {
-            mockInitiateTaskDependencies(CFTTaskState.UNASSIGNED, initiateTaskRequest);
-
-            taskManagementService.initiateTask(taskId, initiateTaskRequest);
-
-            verifyExpectations(CFTTaskState.UNASSIGNED, initiateTaskRequest);
-        }
-
-        @Test
-        void given_initiateTask_without_role_assignment_then_task_is_initiated() {
-            InitiateTaskRequestAttributes initiateRequest = new InitiateTaskRequestAttributes(
-                INITIATION,
-                asList(
-                    new TaskAttribute(TASK_TYPE, A_TASK_TYPE),
-                    new TaskAttribute(TASK_NAME, A_TASK_NAME),
-                    new TaskAttribute(TASK_DUE_DATE, formattedDueDate)
-                )
-            );
-            mockInitiateTaskDependencies(CFTTaskState.UNASSIGNED, initiateRequest);
-            taskManagementService.initiateTask(taskId, initiateRequest);
-
-            verifyExpectations(CFTTaskState.UNASSIGNED, initiateRequest);
-        }
-
-        private void verifyExpectations(CFTTaskState cftTaskState, InitiateTaskRequestAttributes initiateRequest) {
-            verify(cftTaskMapper, atLeastOnce()).mapToTaskResource(taskId, initiateRequest.getTaskAttributes());
-            Map<String, Object> taskAttributes = new HashMap<>();
-
-            if (initiateRequest.getTaskAttributes().stream().anyMatch(t -> t.getName() == TASK_ROLE_ASSIGNMENT_ID)) {
-                taskAttributes.put("roleAssignmentId", SOME_ROLE_ASSIGNMENT_ID);
-            }
-
-            verify(configureTaskService).configureCFTTask(
-                eq(taskResource),
-                ArgumentMatchers.argThat((taskToConfigure) -> taskToConfigure.equals(new TaskToConfigure(
-                    taskId,
-                    A_TASK_TYPE,
-                    "aCaseId",
-                    A_TASK_NAME,
-                    taskAttributes
-                )))
-            );
-
-            verify(taskAutoAssignmentService).autoAssignCFTTask(taskResource);
-
-            if (cftTaskState.equals(CFTTaskState.ASSIGNED) || cftTaskState.equals(CFTTaskState.UNASSIGNED)) {
-                verify(camundaService).updateCftTaskState(
-                    taskId,
-                    cftTaskState.equals(CFTTaskState.ASSIGNED) ? TaskState.ASSIGNED : TaskState.UNASSIGNED
-                );
-            } else {
-                verifyNoInteractions(camundaService);
-            }
-
-            verify(cftTaskDatabaseService).saveTask(taskResource);
-        }
-
-        private void mockInitiateTaskDependencies(CFTTaskState cftTaskState,
-                                                  InitiateTaskRequestAttributes initiateRequest) {
-            when(cftTaskMapper.readDate(any(), any(TaskAttributeDefinition.class), any())).thenCallRealMethod();
-
-            when(cftTaskMapper.mapToTaskResource(taskId, initiateRequest.getTaskAttributes()))
-                .thenReturn(taskResource);
-
-            when(cftTaskMapper.getTaskAttributes(any(TaskResource.class))).thenReturn(taskAttributes);
-
-            when(taskResource.getTaskType()).thenReturn(A_TASK_TYPE);
-            when(taskResource.getTaskId()).thenReturn(taskId);
-            when(taskResource.getCaseId()).thenReturn("aCaseId");
-            when(taskResource.getTaskName()).thenReturn(A_TASK_NAME);
-            when(taskResource.getState()).thenReturn(cftTaskState);
-
-            lenient().when(configureTaskService.configureCFTTask(any(TaskResource.class), any(TaskToConfigure.class)))
-                .thenReturn(taskResource);
-
-            when(taskAutoAssignmentService.autoAssignCFTTask(any(TaskResource.class))).thenReturn(taskResource);
-
-            when(cftTaskDatabaseService.saveTask(any(TaskResource.class))).thenReturn(taskResource);
-        }
-
     }
 
     @Nested
