+++ src/test/java/uk/gov/hmcts/reform/wataskmanagementapi/services/taskmanagementservicetests/InitiateTaskTest.java
@@ -1,6 +1,7 @@
 package uk.gov.hmcts.reform.wataskmanagementapi.services.taskmanagementservicetests;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -11,9 +12,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.reform.wataskmanagementapi.cft.entities.TaskResource;
 import uk.gov.hmcts.reform.wataskmanagementapi.cft.query.CftQueryService;
 import uk.gov.hmcts.reform.wataskmanagementapi.config.LaunchDarklyFeatureFlagProvider;
-import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.InitiateTaskRequestAttributes;
-import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.entities.TaskAttribute;
-import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition;
+import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.InitiateTaskRequestMap;
+import uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition;
 import uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.TaskState;
 import uk.gov.hmcts.reform.wataskmanagementapi.exceptions.v2.DatabaseConflictException;
 import uk.gov.hmcts.reform.wataskmanagementapi.exceptions.v2.validation.CustomConstraintViolationException;
@@ -29,8 +29,9 @@ import uk.gov.hmcts.reform.wataskmanagementapi.taskconfiguration.services.Config
 import uk.gov.hmcts.reform.wataskmanagementapi.taskconfiguration.services.TaskAutoAssignmentService;
 
 import java.sql.SQLException;
+import java.time.OffsetDateTime;
 import java.time.ZonedDateTime;
-import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -39,7 +40,6 @@ import javax.persistence.EntityManager;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
@@ -53,11 +53,10 @@ import static uk.gov.hmcts.reform.wataskmanagementapi.cft.enums.CFTTaskState.ASS
 import static uk.gov.hmcts.reform.wataskmanagementapi.cft.enums.CFTTaskState.UNASSIGNED;
 import static uk.gov.hmcts.reform.wataskmanagementapi.cft.enums.CFTTaskState.UNCONFIGURED;
 import static uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.InitiateTaskOperation.INITIATION;
-import static uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition.TASK_DUE_DATE;
-import static uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition.TASK_LOCATION;
 import static uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition.TASK_NAME;
 import static uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.enums.TaskAttributeDefinition.TASK_TYPE;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaTime.CAMUNDA_DATA_TIME_FORMATTER;
+import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaVariableDefinition.DUE_DATE;
 
 @ExtendWith(MockitoExtension.class)
 @SuppressWarnings("unchecked")
@@ -80,14 +79,12 @@ class InitiateTaskTest extends CamundaHelpers {
     ConfigureTaskService configureTaskService;
     @Mock
     TaskAutoAssignmentService taskAutoAssignmentService;
-    @Mock
-    private Map<String, Object> taskAttributes;
 
     RoleAssignmentVerificationService roleAssignmentVerification;
     TaskManagementService taskManagementService;
     String taskId;
     TaskResource taskResource;
-    private InitiateTaskRequestAttributes initiateTaskRequest;
+    private InitiateTaskRequestMap initiateTaskRequest;
     @Mock
     private EntityManager entityManager;
 
@@ -124,18 +121,19 @@ class InitiateTaskTest extends CamundaHelpers {
             CASE_ID
         );
 
-        List<TaskAttribute> taskAttributeList = new ArrayList<>();
-        taskAttributeList.add(new TaskAttribute(TASK_TYPE, A_TASK_TYPE));
-        taskAttributeList.add(new TaskAttribute(TASK_NAME, A_TASK_NAME));
-        initiateTaskRequest = new InitiateTaskRequestAttributes(INITIATION, taskAttributeList);
+        Map<String, Object> taskAttributes = new HashMap<>();
+        taskAttributes.put(TASK_TYPE.value(), A_TASK_TYPE);
+        taskAttributes.put(TASK_NAME.value(), A_TASK_NAME);
+
+        initiateTaskRequest = new InitiateTaskRequestMap(INITIATION, taskAttributes);
     }
 
     @Test
     void given_initiateTask_then_task_is_saved() {
-        ZonedDateTime dueDate = ZonedDateTime.now();
+        OffsetDateTime dueDate = OffsetDateTime.now();
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
-        initiateTaskRequest.getTaskAttributes().add(new TaskAttribute(TASK_DUE_DATE, formattedDueDate));
-        mockInitiateTaskDependencies();
+        initiateTaskRequest.getTaskAttributes().put(DUE_DATE.value(), formattedDueDate);
+        initiateTaskRequest.getTaskAttributes().put("taskId", taskId);
 
         TaskResource unassignedTaskResource = new TaskResource(
             taskId,
@@ -144,15 +142,16 @@ class InitiateTaskTest extends CamundaHelpers {
             UNASSIGNED,
             CASE_ID
         );
+        unassignedTaskResource.setDueDateTime(dueDate);
 
+        mockInitiateTaskDependencies(unassignedTaskResource);
         when(taskAutoAssignmentService.autoAssignCFTTask(any())).thenReturn(unassignedTaskResource);
-        when(cftTaskMapper.readDate(any(), any(TaskAttributeDefinition.class), any())).thenCallRealMethod();
-        when(cftTaskMapper.getTaskAttributes(any(TaskResource.class))).thenReturn(taskAttributes);
 
+        lenient().when(cftTaskMapper.readDate(any(), any(CamundaVariableDefinition.class), any())).thenCallRealMethod();
         taskManagementService.initiateTask(taskId, initiateTaskRequest);
 
         verify(cftTaskMapper, atLeastOnce()).mapToTaskResource(taskId, initiateTaskRequest.getTaskAttributes());
-        verify(cftTaskMapper, atLeastOnce()).getTaskAttributes(taskResource);
+
         verify(configureTaskService).configureCFTTask(
             eq(taskResource),
             ArgumentMatchers.argThat((taskToConfigure) -> taskToConfigure.equals(new TaskToConfigure(
@@ -160,54 +159,7 @@ class InitiateTaskTest extends CamundaHelpers {
                 A_TASK_TYPE,
                 CASE_ID,
                 A_TASK_NAME,
-                taskAttributes
-            )))
-        );
-
-
-        verify(taskAutoAssignmentService).autoAssignCFTTask(taskResource);
-
-        verify(camundaService).updateCftTaskState(
-            taskId,
-            TaskState.UNASSIGNED
-        );
-
-        verify(cftTaskDatabaseService).saveTask(taskResource);
-    }
-
-    @Test
-    void should_succeed_and_filter_out_nulls() {
-        ZonedDateTime dueDate = ZonedDateTime.now();
-        String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
-        initiateTaskRequest.getTaskAttributes().add(new TaskAttribute(TASK_DUE_DATE, formattedDueDate));
-        mockInitiateTaskDependencies();
-
-        TaskResource unassignedTaskResource = new TaskResource(
-            taskId,
-            A_TASK_NAME,
-            A_TASK_TYPE,
-            UNASSIGNED,
-            CASE_ID
-        );
-
-        when(taskAutoAssignmentService.autoAssignCFTTask(any())).thenReturn(unassignedTaskResource);
-        when(cftTaskMapper.readDate(any(), any(TaskAttributeDefinition.class), any())).thenCallRealMethod();
-        initiateTaskRequest.getTaskAttributes().add(new TaskAttribute(TASK_LOCATION, null));
-        when(cftTaskMapper.getTaskAttributes(any(TaskResource.class))).thenReturn(taskAttributes);
-
-        TaskResource taskResource = taskManagementService.initiateTask(taskId, initiateTaskRequest);
-
-        assertNull(taskResource.getLocation());
-        verify(cftTaskMapper, atLeastOnce()).mapToTaskResource(taskId, initiateTaskRequest.getTaskAttributes());
-        verify(cftTaskMapper).getTaskAttributes(taskResource);
-        verify(configureTaskService).configureCFTTask(
-            eq(taskResource),
-            ArgumentMatchers.argThat((taskToConfigure) -> taskToConfigure.equals(new TaskToConfigure(
-                taskId,
-                A_TASK_TYPE,
-                CASE_ID,
-                A_TASK_NAME,
-                taskAttributes
+                getTaskAttributesWithDueDateUpdate(dueDate)
             )))
         );
 
@@ -224,10 +176,10 @@ class InitiateTaskTest extends CamundaHelpers {
 
     @Test
     void given_initiateTask_with_previous_valid_assignee_should_keep_assignee() {
-        ZonedDateTime dueDate = ZonedDateTime.now();
+        OffsetDateTime dueDate = OffsetDateTime.now();
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
-        initiateTaskRequest.getTaskAttributes().add(new TaskAttribute(TASK_DUE_DATE, formattedDueDate));
-        mockInitiateTaskDependencies();
+        initiateTaskRequest.getTaskAttributes().put(DUE_DATE.value(), formattedDueDate);
+        initiateTaskRequest.getTaskAttributes().put("taskId", taskId);
 
         TaskResource taskWithAssignee = new TaskResource(
             taskId,
@@ -238,11 +190,16 @@ class InitiateTaskTest extends CamundaHelpers {
             "someUserId"
         );
 
+        taskWithAssignee.setDueDateTime(dueDate);
+        mockInitiateTaskDependencies(taskWithAssignee);
+
         when(configureTaskService.configureCFTTask(any(), any())).thenReturn(taskWithAssignee);
         when(taskAutoAssignmentService.checkAssigneeIsStillValid(any(), eq("someUserId"))).thenReturn(true);
 
-        when(cftTaskMapper.readDate(any(), any(TaskAttributeDefinition.class), any())).thenCallRealMethod();
-        when(cftTaskMapper.getTaskAttributes(any(TaskResource.class))).thenReturn(taskAttributes);
+        when(cftTaskMapper.readDate(any(), any(CamundaVariableDefinition.class), any())).thenCallRealMethod();
+
+        when(cftTaskMapper.mapToTaskResource(taskId, initiateTaskRequest.getTaskAttributes()))
+            .thenReturn(taskWithAssignee);
 
         TaskResource taskResource = taskManagementService.initiateTask(taskId, initiateTaskRequest);
 
@@ -255,7 +212,7 @@ class InitiateTaskTest extends CamundaHelpers {
                 A_TASK_TYPE,
                 CASE_ID,
                 A_TASK_NAME,
-                taskAttributes
+                getTaskAttributesWithDueDateUpdate(dueDate)
             )))
         );
 
@@ -267,12 +224,19 @@ class InitiateTaskTest extends CamundaHelpers {
         verify(cftTaskDatabaseService).saveTask(taskResource);
     }
 
+    @NotNull
+    private Map<String, Object> getTaskAttributesWithDueDateUpdate(OffsetDateTime dueDate) {
+        Map<String, Object> taskAttributes = new HashMap<>(initiateTaskRequest.getTaskAttributes());
+        taskAttributes.put(DUE_DATE.value(), dueDate);
+        return taskAttributes;
+    }
+
     @Test
     void given_initiateTask_with_previous_invalid_assignee_should_reassign() {
-        ZonedDateTime dueDate = ZonedDateTime.now();
+        OffsetDateTime dueDate = OffsetDateTime.now();
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
-        initiateTaskRequest.getTaskAttributes().add(new TaskAttribute(TASK_DUE_DATE, formattedDueDate));
-        mockInitiateTaskDependencies();
+        initiateTaskRequest.getTaskAttributes().put(DUE_DATE.value(), formattedDueDate);
+        initiateTaskRequest.getTaskAttributes().put("taskId", taskId);
 
         TaskResource taskWithAssignee = new TaskResource(
             taskId,
@@ -282,6 +246,7 @@ class InitiateTaskTest extends CamundaHelpers {
             CASE_ID,
             "someUserId"
         );
+        mockInitiateTaskDependencies(taskWithAssignee);
 
 
         TaskResource taskReassigned = new TaskResource(
@@ -292,19 +257,21 @@ class InitiateTaskTest extends CamundaHelpers {
             CASE_ID,
             "anotherUserId"
         );
+        taskWithAssignee.setDueDateTime(dueDate);
+        taskReassigned.setDueDateTime(dueDate);
 
         when(configureTaskService.configureCFTTask(any(), any())).thenReturn(taskWithAssignee);
         when(taskAutoAssignmentService.checkAssigneeIsStillValid(any(), eq("someUserId"))).thenReturn(false);
 
         when(taskAutoAssignmentService.autoAssignCFTTask(any())).thenReturn(taskReassigned);
 
-        when(cftTaskMapper.readDate(any(), any(TaskAttributeDefinition.class), any())).thenCallRealMethod();
-
-        when(cftTaskMapper.getTaskAttributes(any(TaskResource.class))).thenReturn(taskAttributes);
+        lenient().when(cftTaskMapper.readDate(any(), any(CamundaVariableDefinition.class), any())).thenCallRealMethod();
 
         TaskResource taskResource = taskManagementService.initiateTask(taskId, initiateTaskRequest);
 
         verify(cftTaskMapper, atLeastOnce()).mapToTaskResource(taskId, initiateTaskRequest.getTaskAttributes());
+
+        Map<String, Object> taskAttributes = getTaskAttributesWithDueDateUpdate(dueDate);
         verify(configureTaskService).configureCFTTask(
             eq(taskResource),
             ArgumentMatchers.argThat((taskToConfigure) -> taskToConfigure.equals(new TaskToConfigure(
@@ -342,7 +309,7 @@ class InitiateTaskTest extends CamundaHelpers {
             .when(cftTaskDatabaseService).insertAndLock(anyString(), any());
         ZonedDateTime dueDate = ZonedDateTime.now();
         String formattedDueDate = CAMUNDA_DATA_TIME_FORMATTER.format(dueDate);
-        initiateTaskRequest.getTaskAttributes().add(new TaskAttribute(TASK_DUE_DATE, formattedDueDate));
+        initiateTaskRequest.getTaskAttributes().put(DUE_DATE.value(), formattedDueDate);
 
         assertThatThrownBy(() -> taskManagementService.initiateTask(taskId, initiateTaskRequest)
         )
@@ -355,29 +322,23 @@ class InitiateTaskTest extends CamundaHelpers {
     @Test
     void should_set_task_attributes_when_initiate_task_request_initiated() {
 
-        List<TaskAttribute> taskAttributes = initiateTaskRequest.getTaskAttributes();
+        Map<String, Object> taskAttributes = initiateTaskRequest.getTaskAttributes();
 
         assertNotNull(taskAttributes);
-        assertNotNull(taskAttributes.get(0));
-        assertEquals(TASK_TYPE, taskAttributes.get(0).getName());
-        assertEquals(A_TASK_TYPE, taskAttributes.get(0).getValue());
-
-        assertNotNull(taskAttributes.get(1));
-        assertEquals(TASK_NAME, taskAttributes.get(1).getName());
-        assertEquals(A_TASK_NAME, taskAttributes.get(1).getValue());
-
+        assertEquals(A_TASK_TYPE, taskAttributes.get(TASK_TYPE.value()));
+        assertEquals(A_TASK_NAME, taskAttributes.get(TASK_NAME.value()));
     }
 
-    private void mockInitiateTaskDependencies() {
+    private void mockInitiateTaskDependencies(TaskResource expected) {
         when(cftTaskMapper.mapToTaskResource(taskId, initiateTaskRequest.getTaskAttributes()))
-            .thenReturn(taskResource);
+            .thenReturn(expected);
 
         lenient().when(configureTaskService.configureCFTTask(any(TaskResource.class), any(TaskToConfigure.class)))
-            .thenReturn(taskResource);
+            .thenReturn(expected);
 
-        lenient().when(taskAutoAssignmentService.autoAssignCFTTask(any(TaskResource.class))).thenReturn(taskResource);
+        lenient().when(taskAutoAssignmentService.autoAssignCFTTask(any(TaskResource.class))).thenReturn(expected);
 
-        lenient().when(cftTaskDatabaseService.saveTask(any(TaskResource.class))).thenReturn(taskResource);
+        lenient().when(cftTaskDatabaseService.saveTask(any(TaskResource.class))).thenReturn(expected);
     }
 }
